Finish frontend auth managament

+++ server/controllers/taskController.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler");
 const Task = require("../models/taskModel");
 
 const getTasks = asyncHandler(async (req, res) => {
-  const list = await Task.find();
+  const list = await Task.find({ user_id: req.user.id });
   res.status(200).json(list);
 });
 
@@ -14,6 +14,7 @@ const createTask = asyncHandler(async (req, res) => {
     throw new Error("All fields are mandatory! ");
   }
   const list = await Task.create({
+    user_id: req.user.id,
     //had to change task to Task because i have already declared it in const{task}=req.body
     task,
   });
@@ -25,6 +26,11 @@ const updateTask = asyncHandler(async (req, res) => {
   if (!list) {
     res.status(404);
     throw new Error("Task was not found");
+  }
+
+  if (list.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User doesn't have the permission to update this contact");
   }
   const updatedList = await Task.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
@@ -47,6 +53,10 @@ const deleteTask = asyncHandler(async (req, res) => {
   if (!list) {
     res.status(404);
     throw new Error("Task was not found");
+  }
+  if (list.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User doesn't have the permission to delete this contact");
   }
   await Task.deleteOne({ _id: req.params.id });
   res.status(200).json(list);
